+++ jax/_src/interpreters/xla.py
@@ -502,3 +502,9 @@ class _BackendSpecificTranslationsAdapter(defaultdict):
 
 backend_specific_translations: Dict[str, _TranslationRuleAdapter]
 backend_specific_translations = _BackendSpecificTranslationsAdapter()
+
+
+# TODO(yashkatariya): Delete this.
+def device_put(x, device=None):
+  from jax._src import api
+  return api.device_put(x, device)