+++ jax/_src/test_util.py
@@ -178,7 +178,6 @@ def capture_stdout() -> Generator[Callable[[], str], None, None]:
 
 @contextmanager
 def count_device_put():
-  device_put = dispatch.device_put
   batched_device_put = pxla.batched_device_put
   count = [0]
 
@@ -189,24 +188,19 @@ def count_device_put():
       # underlying payload or several). We only want to count these
       # recursive puts once, so we skip counting more than the outermost
       # one in such a call stack.
-      dispatch.device_put = device_put
       pxla.batched_device_put = batched_device_put
       try:
         return fn(*args, **kwargs)
       finally:
-        dispatch.device_put = device_put_and_count
         pxla.batched_device_put = batched_device_put_and_count
     return fn_and_count
 
-  device_put_and_count = make_fn_and_count(device_put)
   batched_device_put_and_count = make_fn_and_count(batched_device_put)
 
-  dispatch.device_put = device_put_and_count
   pxla.batched_device_put = batched_device_put_and_count
   try:
     yield count
   finally:
-    dispatch.device_put = device_put
     pxla.batched_device_put = batched_device_put