+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Remember to align the itemized text with the first line of an item within a list
     for which it is an alias.
   * Passing additional arguments to {func}`jax.numpy.ndarray.at` by position is deprecated.
     For example, instead of `x.at[i].get(True)`, use `x.at[i].get(indices_are_sorted=True)`
+  * `jax.interpreters.xla.device_put` is deprecated. Please use `jax.device_put`.
+  * `jax.interpreters.pxla.device_put` is deprecated. Please use `jax.device_put`.
 
 ## jaxlib 0.4.7